requirements.txt
matplotlib
pandas_datareader

#new
# ecos
# cvxpy

#NEW
Flask
# heroku needs this to run our app on the server
gunicorn
python-dotenv
PyPortfolioOpt

#NEW AGAIN
mpld3

#BOKEH integration
bokeh

